Improve structure of likelihoods subdirectory (#1416)
Addresses #1405 

* New structure underneath gpflow/likelihoods/:
  * base.py: all base classes (Likelihood, MonteCarloLikelihood, ScalarLikelihood) and SwitchedLikelihood
  * multiclass.py: multi-class classification (Softmax, MultiClass + RobustMax)
  * scalar_continuous.py: continuous-Y subclasses of ScalarLikelihood (Gaussian, StudentT, Exponential, Beta, Gamma)
  * scalar_discrete.py: discrete-Y subclasses of ScalarLikelihood (Bernoulli, Poisson, Ordinal)
  * utils.py: the `inv_probit` link function used by Bernoulli and Beta likelihoods
  * misc.py: GaussianMC - used for demonstration/tests only.
  (Note that usage, i.e. accessing gpflow.likelihoods.<LikelihoodClass>, has not changed.)

* Tests for multi-class classification likelihoods moved out into their own test module (including stubs for the missing MultiClass quadrature tests of #1091)

* Re-activates the quadrature tests for ScalarLikelihood subclasses with analytic variational_expectations/predict_log_density/predict_mean_and_var that inadvertently got disabled by #1334 

* Fixes a bug in Bernoulli._predict_log_density that was uncovered by these tests

* Fixes random seed for mock data generation in test_natural_gradient to make svgp_vs_gpr test pass again
