+++ gpflow/likelihoods/__init__.py
@@ -1,18 +1,19 @@
-from .likelihoods import (
+from .base import Likelihood, ScalarLikelihood, SwitchedLikelihood, MonteCarloLikelihood
+from .scalar_discrete import (
     Bernoulli,
+    Ordinal,
+    Poisson,
+)
+from .scalar_continuous import (
     Beta,
     Exponential,
     Gamma,
     Gaussian,
-    GaussianMC,
-    Likelihood,
-    MonteCarloLikelihood,
-    MultiClass,
-    Ordinal,
-    Poisson,
-    Softmax,
-    SwitchedLikelihood,
     StudentT,
+)
+from .misc import GaussianMC
+from .multiclass import (
+    MultiClass,
+    Softmax,
     RobustMax,
-    ScalarLikelihood,
 )
