+++ gpflow/likelihoods/base.py
@@ -57,17 +57,8 @@ import tensorflow as tf
 import abc
 import warnings
 
-from .. import logdensities
-from ..base import Module, Parameter
-from ..config import default_float
+from ..base import Module
 from ..quadrature import hermgauss, ndiag_mc, ndiagquad
-from ..utilities import positive, to_default_int
-from .robustmax import RobustMax
-
-
-def inv_probit(x):
-    jitter = 1e-3  # ensures output is strictly between 0 and 1
-    return 0.5 * (1.0 + tf.math.erf(x / np.sqrt(2.0))) * (1 - 2 * jitter) + jitter
 
 
 class Likelihood(Module, metaclass=abc.ABCMeta):
@@ -389,285 +380,6 @@ class ScalarLikelihood(Likelihood):
         return E_y, V_y
 
 
-class Gaussian(ScalarLikelihood):
-    r"""
-    The Gaussian likelihood is appropriate where uncertainties associated with the data are
-    believed to follow a normal distribution, with constant variance.
-
-    Very small uncertainties can lead to numerical instability during the
-    optimization process. A lower bound of 1e-6 is therefore imposed on the likelihood variance
-    by default.
-    """
-
-    def __init__(self, variance=1.0, variance_lower_bound=1e-6, **kwargs):
-        super().__init__(**kwargs)
-        self.variance = Parameter(variance, transform=positive(lower=variance_lower_bound))
-
-    def _scalar_log_prob(self, F, Y):
-        return logdensities.gaussian(Y, F, self.variance)
-
-    def _conditional_mean(self, F):  # pylint: disable=R0201
-        return tf.identity(F)
-
-    def _conditional_variance(self, F):
-        return tf.fill(tf.shape(F), tf.squeeze(self.variance))
-
-    def _predict_mean_and_var(self, Fmu, Fvar):
-        return tf.identity(Fmu), Fvar + self.variance
-
-    def _predict_log_density(self, Fmu, Fvar, Y):
-        return tf.reduce_sum(logdensities.gaussian(Y, Fmu, Fvar + self.variance), axis=-1)
-
-    def _variational_expectations(self, Fmu, Fvar, Y):
-        return tf.reduce_sum(
-            -0.5 * np.log(2 * np.pi)
-            - 0.5 * tf.math.log(self.variance)
-            - 0.5 * ((Y - Fmu) ** 2 + Fvar) / self.variance,
-            axis=-1,
-        )
-
-
-class Poisson(ScalarLikelihood):
-    r"""
-    Poisson likelihood for use with count data, where the rate is given by the (transformed) GP.
-
-    let g(.) be the inverse-link function, then this likelihood represents
-
-    p(yᵢ | fᵢ) = Poisson(yᵢ | g(fᵢ) * binsize)
-
-    Note:binsize
-    For use in a Log Gaussian Cox process (doubly stochastic model) where the
-    rate function of an inhomogeneous Poisson process is given by a GP.  The
-    intractable likelihood can be approximated via a Riemann sum (with bins
-    of size 'binsize') and using this Poisson likelihood.
-    """
-
-    def __init__(self, invlink=tf.exp, binsize=1.0, **kwargs):
-        super().__init__(**kwargs)
-        self.invlink = invlink
-        self.binsize = np.array(binsize, dtype=default_float())
-
-    def _scalar_log_prob(self, F, Y):
-        return logdensities.poisson(Y, self.invlink(F) * self.binsize)
-
-    def _conditional_variance(self, F):
-        return self.invlink(F) * self.binsize
-
-    def _conditional_mean(self, F):
-        return self.invlink(F) * self.binsize
-
-    def _variational_expectations(self, Fmu, Fvar, Y):
-        if self.invlink is tf.exp:
-            return tf.reduce_sum(
-                Y * Fmu
-                - tf.exp(Fmu + Fvar / 2) * self.binsize
-                - tf.math.lgamma(Y + 1)
-                + Y * tf.math.log(self.binsize),
-                axis=-1,
-            )
-        return super()._variational_expectations(Fmu, Fvar, Y)
-
-
-class Exponential(ScalarLikelihood):
-    def __init__(self, invlink=tf.exp, **kwargs):
-        super().__init__(**kwargs)
-        self.invlink = invlink
-
-    def _scalar_log_prob(self, F, Y):
-        return logdensities.exponential(Y, self.invlink(F))
-
-    def _conditional_mean(self, F):
-        return self.invlink(F)
-
-    def _conditional_variance(self, F):
-        return tf.square(self.invlink(F))
-
-    def _variational_expectations(self, Fmu, Fvar, Y):
-        if self.invlink is tf.exp:
-            return tf.reduce_sum(-tf.exp(-Fmu + Fvar / 2) * Y - Fmu, axis=-1)
-        return super()._variational_expectations(Fmu, Fvar, Y)
-
-
-class StudentT(ScalarLikelihood):
-    def __init__(self, scale=1.0, df=3.0, **kwargs):
-        """
-        :param scale float: scale parameter
-        :param df float: degrees of freedom
-        """
-        super().__init__(**kwargs)
-        self.df = df
-        self.scale = Parameter(scale, transform=positive())
-
-    def _scalar_log_prob(self, F, Y):
-        return logdensities.student_t(Y, F, self.scale, self.df)
-
-    def _conditional_mean(self, F):
-        return F
-
-    def _conditional_variance(self, F):
-        var = (self.scale ** 2) * (self.df / (self.df - 2.0))
-        return tf.fill(tf.shape(F), tf.squeeze(var))
-
-
-class Bernoulli(ScalarLikelihood):
-    def __init__(self, invlink=inv_probit, **kwargs):
-        super().__init__(**kwargs)
-        self.invlink = invlink
-
-    def _scalar_log_prob(self, F, Y):
-        return logdensities.bernoulli(Y, self.invlink(F))
-
-    def _predict_mean_and_var(self, Fmu, Fvar):
-        if self.invlink is inv_probit:
-            p = inv_probit(Fmu / tf.sqrt(1 + Fvar))
-            return p, p - tf.square(p)
-        else:
-            # for other invlink, use quadrature
-            return super()._predict_mean_and_var(Fmu, Fvar)
-
-    def _predict_log_density(self, Fmu, Fvar, Y):
-        p = self.predict_mean_and_var(Fmu, Fvar)[0]
-        return logdensities.bernoulli(Y, p)
-
-    def _conditional_mean(self, F):
-        return self.invlink(F)
-
-    def _conditional_variance(self, F):
-        p = self.conditional_mean(F)
-        return p - (p ** 2)
-
-
-class Gamma(ScalarLikelihood):
-    """
-    Use the transformed GP to give the *scale* (inverse rate) of the Gamma
-    """
-
-    def __init__(self, invlink=tf.exp, **kwargs):
-        super().__init__(**kwargs)
-        self.invlink = invlink
-        self.shape = Parameter(1.0, transform=positive())
-
-    def _scalar_log_prob(self, F, Y):
-        return logdensities.gamma(Y, self.shape, self.invlink(F))
-
-    def _conditional_mean(self, F):
-        return self.shape * self.invlink(F)
-
-    def _conditional_variance(self, F):
-        scale = self.invlink(F)
-        return self.shape * (scale ** 2)
-
-    def _variational_expectations(self, Fmu, Fvar, Y):
-        if self.invlink is tf.exp:
-            return tf.reduce_sum(
-                -self.shape * Fmu
-                - tf.math.lgamma(self.shape)
-                + (self.shape - 1.0) * tf.math.log(Y)
-                - Y * tf.exp(-Fmu + Fvar / 2.0),
-                axis=-1,
-            )
-        else:
-            return super()._variational_expectations(Fmu, Fvar, Y)
-
-
-class Beta(ScalarLikelihood):
-    """
-    This uses a reparameterisation of the Beta density. We have the mean of the
-    Beta distribution given by the transformed process:
-
-        m = invlink(f)
-
-    and a scale parameter. The familiar α, β parameters are given by
-
-        m     = α / (α + β)
-        scale = α + β
-
-    so:
-        α = scale * m
-        β  = scale * (1-m)
-    """
-
-    def __init__(self, invlink=inv_probit, scale=1.0, **kwargs):
-        super().__init__(**kwargs)
-        self.scale = Parameter(scale, transform=positive())
-        self.invlink = invlink
-
-    def _scalar_log_prob(self, F, Y):
-        mean = self.invlink(F)
-        alpha = mean * self.scale
-        beta = self.scale - alpha
-        return logdensities.beta(Y, alpha, beta)
-
-    def _conditional_mean(self, F):
-        return self.invlink(F)
-
-    def _conditional_variance(self, F):
-        mean = self.invlink(F)
-        return (mean - tf.square(mean)) / (self.scale + 1.0)
-
-
-class MultiClass(Likelihood):
-    def __init__(self, num_classes, invlink=None, **kwargs):
-        """
-        A likelihood for multi-way classification.  Currently the only valid
-        choice of inverse-link function (invlink) is an instance of RobustMax.
-
-        For most problems, the stochastic `Softmax` likelihood may be more
-        appropriate (note that you then cannot use Scipy optimizer).
-        """
-        super().__init__(latent_dim=num_classes, observation_dim=None, **kwargs)
-        self.num_classes = num_classes
-        self.num_gauss_hermite_points = 20
-
-        if invlink is None:
-            invlink = RobustMax(self.num_classes)
-
-        if not isinstance(invlink, RobustMax):
-            raise NotImplementedError
-
-        self.invlink = invlink
-
-    def _log_prob(self, F, Y):
-        hits = tf.equal(tf.expand_dims(tf.argmax(F, 1), 1), tf.cast(Y, tf.int64))
-        yes = tf.ones(tf.shape(Y), dtype=default_float()) - self.invlink.epsilon
-        no = tf.zeros(tf.shape(Y), dtype=default_float()) + self.invlink.eps_k1
-        p = tf.where(hits, yes, no)
-        return tf.reduce_sum(tf.math.log(p), axis=-1)
-
-    def _variational_expectations(self, Fmu, Fvar, Y):
-        gh_x, gh_w = hermgauss(self.num_gauss_hermite_points)
-        p = self.invlink.prob_is_largest(Y, Fmu, Fvar, gh_x, gh_w)
-        ve = p * tf.math.log(1.0 - self.invlink.epsilon) + (1.0 - p) * tf.math.log(
-            self.invlink.eps_k1
-        )
-        return tf.reduce_sum(ve, axis=-1)
-
-    def _predict_mean_and_var(self, Fmu, Fvar):
-        possible_outputs = [
-            tf.fill(tf.stack([tf.shape(Fmu)[0], 1]), np.array(i, dtype=np.int64))
-            for i in range(self.num_classes)
-        ]
-        ps = [self._predict_non_logged_density(Fmu, Fvar, po) for po in possible_outputs]
-        ps = tf.transpose(tf.stack([tf.reshape(p, (-1,)) for p in ps]))
-        return ps, ps - tf.square(ps)
-
-    def _predict_log_density(self, Fmu, Fvar, Y):
-        return tf.reduce_sum(tf.math.log(self._predict_non_logged_density(Fmu, Fvar, Y)), axis=-1)
-
-    def _predict_non_logged_density(self, Fmu, Fvar, Y):
-        gh_x, gh_w = hermgauss(self.num_gauss_hermite_points)
-        p = self.invlink.prob_is_largest(Y, Fmu, Fvar, gh_x, gh_w)
-        den = p * (1.0 - self.invlink.epsilon) + (1.0 - p) * (self.invlink.eps_k1)
-        return den
-
-    def _conditional_mean(self, F):
-        return self.invlink(F)
-
-    def _conditional_variance(self, F):
-        p = self.conditional_mean(F)
-        return p - tf.square(p)
-
-
 class SwitchedLikelihood(ScalarLikelihood):
     def __init__(self, likelihood_list, **kwargs):
         """
@@ -732,85 +444,6 @@ class SwitchedLikelihood(ScalarLikelihood):
 
     def _conditional_variance(self, F):
         raise NotImplementedError
-
-
-class Ordinal(ScalarLikelihood):
-    """
-    A likelihood for doing ordinal regression.
-
-    The data are integer values from 0 to k, and the user must specify (k-1)
-    'bin edges' which define the points at which the labels switch. Let the bin
-    edges be [a₀, a₁, ... aₖ₋₁], then the likelihood is
-
-    p(Y=0|F) = ɸ((a₀ - F) / σ)
-    p(Y=1|F) = ɸ((a₁ - F) / σ) - ɸ((a₀ - F) / σ)
-    p(Y=2|F) = ɸ((a₂ - F) / σ) - ɸ((a₁ - F) / σ)
-    ...
-    p(Y=K|F) = 1 - ɸ((aₖ₋₁ - F) / σ)
-
-    where ɸ is the cumulative density function of a Gaussian (the inverse probit
-    function) and σ is a parameter to be learned. A reference is:
-
-    @article{chu2005gaussian,
-      title={Gaussian processes for ordinal regression},
-      author={Chu, Wei and Ghahramani, Zoubin},
-      journal={Journal of Machine Learning Research},
-      volume={6},
-      number={Jul},
-      pages={1019--1041},
-      year={2005}
-    }
-    """
-
-    def __init__(self, bin_edges, **kwargs):
-        """
-        bin_edges is a numpy array specifying at which function value the
-        output label should switch. If the possible Y values are 0...K, then
-        the size of bin_edges should be (K-1).
-        """
-        super().__init__(**kwargs)
-        self.bin_edges = bin_edges
-        self.num_bins = bin_edges.size + 1
-        self.sigma = Parameter(1.0, transform=positive())
-
-    def _scalar_log_prob(self, F, Y):
-        Y = to_default_int(Y)
-        scaled_bins_left = tf.concat([self.bin_edges / self.sigma, np.array([np.inf])], 0)
-        scaled_bins_right = tf.concat([np.array([-np.inf]), self.bin_edges / self.sigma], 0)
-        selected_bins_left = tf.gather(scaled_bins_left, Y)
-        selected_bins_right = tf.gather(scaled_bins_right, Y)
-
-        return tf.math.log(
-            inv_probit(selected_bins_left - F / self.sigma)
-            - inv_probit(selected_bins_right - F / self.sigma)
-            + 1e-6
-        )
-
-    def _make_phi(self, F):
-        """
-        A helper function for making predictions. Constructs a probability
-        matrix where each row output the probability of the corresponding
-        label, and the rows match the entries of F.
-
-        Note that a matrix of F values is flattened.
-        """
-        scaled_bins_left = tf.concat([self.bin_edges / self.sigma, np.array([np.inf])], 0)
-        scaled_bins_right = tf.concat([np.array([-np.inf]), self.bin_edges / self.sigma], 0)
-        return inv_probit(scaled_bins_left - tf.reshape(F, (-1, 1)) / self.sigma) - inv_probit(
-            scaled_bins_right - tf.reshape(F, (-1, 1)) / self.sigma
-        )
-
-    def _conditional_mean(self, F):
-        phi = self._make_phi(F)
-        Ys = tf.reshape(np.arange(self.num_bins, dtype=default_float()), (-1, 1))
-        return tf.reshape(tf.linalg.matmul(phi, Ys), tf.shape(F))
-
-    def _conditional_variance(self, F):
-        phi = self._make_phi(F)
-        Ys = tf.reshape(np.arange(self.num_bins, dtype=default_float()), (-1, 1))
-        E_y = phi @ Ys
-        E_y2 = phi @ (Ys ** 2)
-        return tf.reshape(E_y2 - E_y ** 2, tf.shape(F))
 
 
 class MonteCarloLikelihood(Likelihood):
@@ -895,34 +528,3 @@ class MonteCarloLikelihood(Likelihood):
         return tf.reduce_sum(
             self._mc_quadrature(self.log_prob, Fmu, Fvar, Y=Y, epsilon=epsilon), axis=-1
         )
-
-
-class GaussianMC(MonteCarloLikelihood, Gaussian):
-    """
-    Stochastic version of Gaussian likelihood for comparison.
-    """
-
-    pass
-
-
-class Softmax(MonteCarloLikelihood):
-    """
-    The soft-max multi-class likelihood.  It can only provide a stochastic
-    Monte-Carlo estimate of the variational expectations term, but this
-    added variance tends to be small compared to that due to mini-batching
-    (when using the SVGP model).
-    """
-
-    def __init__(self, num_classes, **kwargs):
-        super().__init__(latent_dim=num_classes, observation_dim=None, **kwargs)
-        self.num_classes = self.latent_dim
-
-    def _log_prob(self, F, Y):
-        return -tf.nn.sparse_softmax_cross_entropy_with_logits(logits=F, labels=Y[:, 0])
-
-    def _conditional_mean(self, F):
-        return tf.nn.softmax(F)
-
-    def _conditional_variance(self, F):
-        p = self.conditional_mean(F)
-        return p - p ** 2
