+++ gpflow/likelihoods/utils.py
@@ -1,0 +1,7 @@
+import numpy as np
+import tensorflow as tf
+
+
+def inv_probit(x):
+    jitter = 1e-3  # ensures output is strictly between 0 and 1
+    return 0.5 * (1.0 + tf.math.erf(x / np.sqrt(2.0))) * (1 - 2 * jitter) + jitter
